+++ packages/core/src/createKubb.test.ts
@@ -6,9 +6,9 @@ import { createFile, createSource, createText } from '@kubb/ast'
 import { afterEach, describe, expect, it, test, vi } from 'vitest'
 import { createMockedAdapter } from '#mocks'
 import { createKubb } from './createKubb.ts'
-import { definePlugin } from './definePlugin.ts'
 import { createPlugin } from './createPlugin.ts'
-import type { Config, KubbHooks, Plugin, UserConfig } from './types.ts'
+import { definePlugin } from './definePlugin.ts'
+import type { Config, KubbHooks, Plugin } from './types.ts'
 
 describe('createKubb', () => {
   const pluginMocks = {
@@ -250,7 +250,7 @@ describe('createKubb', () => {
         }
       })
 
-      const excludeConfig: UserConfig = {
+      const excludeConfig: Config = {
         ...config,
         output: {
           ...config.output,
