+++ packages/core/src/types.ts
@@ -6,68 +6,11 @@ import type { RendererFactory } from './createRenderer.ts'
 import type { Storage } from './createStorage.ts'
 import type { Generator } from './defineGenerator.ts'
 import type { Parser } from './defineParser.ts'
-import type { HookStylePlugin } from './definePlugin.ts'
 import type { KubbHooks } from './Kubb.ts'
 import type { PluginDriver } from './PluginDriver.ts'
 
 export type { Printer, PrinterFactoryOptions, PrinterPartial } from '@kubb/ast/types'
 export type { Renderer, RendererFactory } from './createRenderer.ts'
-
-/**
- * Config used in `kubb.config.ts`
- *
- * @example
- * import { defineConfig } from '@kubb/core'
- * export default defineConfig({
- * ...
- * })
- */
-export type UserConfig<TInput = Input> = Omit<Config<TInput>, 'root' | 'plugins' | 'parsers' | 'adapter'> & {
-  /**
-   * The project root directory, which can be either an absolute path or a path relative to the location of your `kubb.config.ts` file.
-   * @default process.cwd()
-   */
-  root?: string
-  /**
-   * An array of parsers used to convert generated files to strings.
-   * Each parser handles specific file extensions (e.g. `.ts`, `.tsx`).
-   *
-   * A catch-all fallback parser is always appended last for any unhandled extension.
-   *
-   * When omitted, `parserTsx` from `@kubb/parser-ts` is used automatically as the
-   * default (requires `@kubb/parser-ts` to be installed as an optional dependency).
-   * @default [parserTsx] — from `@kubb/parser-ts`
-   * @example
-   * ```ts
-   * import { parserTs, tsxParser } from '@kubb/parser-ts'
-   * export default defineConfig({
-   *   parsers: [parserTs, tsxParser],
-   * })
-   * ```
-   */
-  parsers?: Array<Parser>
-  /**
-   * Adapter that converts the input file into a `@kubb/ast` `InputNode` — the universal
-   * intermediate representation consumed by all Kubb plugins.
-   *
-   * When omitted, `adapterOas()` from `@kubb/adapter-oas` is used automatically as the
-   * default (requires `@kubb/adapter-oas` to be installed as an optional dependency).
-   *
-   * - Use `@kubb/adapter-oas` for OpenAPI / Swagger (default).
-   * - Use `@kubb/adapter-drizzle` or `@kubb/adapter-asyncapi` for other formats.
-   *
-   * @default adapterOas() — from `@kubb/adapter-oas`
-   */
-  adapter?: Adapter
-  /**
-   * An array of Kubb plugins used for code generation.
-   * Each plugin may declare additional configurable options.
-   * If a plugin depends on another, an error is thrown when the dependency is missing.
-   * Use `dependencies` on the plugin to declare execution order.
-   */
-  // inject needs to be omitted because else we have a clash with the PluginDriver instance
-  plugins?: Array<Omit<UnknownUserPlugin, 'inject'> | HookStylePlugin>
-}
 
 export type InputPath = {
   /**
@@ -531,8 +474,6 @@ export type UserPlugin<TOptions extends PluginFactoryOptions = PluginFactoryOpti
  * @deprecated
  */
 export type UserPluginWithLifeCycle<TOptions extends PluginFactoryOptions = PluginFactoryOptions> = UserPlugin<TOptions> & PluginLifecycle<TOptions>
-
-type UnknownUserPlugin = UserPlugin<PluginFactoryOptions<string, object, object, unknown, object>>
 
 /**
  * Handler for a single schema node. Used by the `schema` hook on a plugin.
@@ -1203,7 +1144,7 @@ export type CLIOptions = {
  * @deprecated
  * TODO: can we remove this?
  */
-export type ConfigInput = PossiblePromise<UserConfig | UserConfig[]> | ((cli: CLIOptions) => PossiblePromise<UserConfig | UserConfig[]>)
+export type ConfigInput = PossiblePromise<Config | Config[]> | ((cli: CLIOptions) => PossiblePromise<Config | Config[]>)
 
 export type { BuildOutput } from './createKubb.ts'
 export type { Parser } from './defineParser.ts'
