+++ packages/kubb/src/index.ts
@@ -1,1 +1,2 @@
+export type { ConfigInput, UserConfig } from './defineConfig.ts'
 export { defineConfig } from './defineConfig.ts'
