+++ packages/mcp/src/utils/resolveUserConfig.ts
@@ -1,5 +1,5 @@
 import { isPromise } from '@internals/utils'
-import type { CLIOptions, Config, UserConfig } from '@kubb/core'
+import type { CLIOptions, Config } from '@kubb/core'
 
 export type ResolveUserConfigOptions = {
   configPath?: string
@@ -9,11 +9,11 @@ export type ResolveUserConfigOptions = {
 /**
  * Resolve the config by handling function configs and returning the final configuration
  */
-export async function resolveUserConfig(userConfig: UserConfig, options: ResolveUserConfigOptions): Promise<Config> {
-  let kubbUserConfig = Promise.resolve(userConfig) as Promise<UserConfig>
+export async function resolveUserConfig(config: Config, options: ResolveUserConfigOptions): Promise<Config> {
+  let kubbUserConfig = Promise.resolve(config) as Promise<Config>
 
-  if (typeof userConfig === 'function') {
-    const possiblePromise = (userConfig as any)({ logLevel: options.logLevel, config: options.configPath } as CLIOptions)
+  if (typeof config === 'function') {
+    const possiblePromise = (config as any)({ logLevel: options.logLevel, config: options.configPath } as CLIOptions)
     if (isPromise(possiblePromise)) {
       kubbUserConfig = possiblePromise
     } else {
