+++ packages/unplugin-kubb/src/types.ts
@@ -1,8 +1,8 @@
-import type { UserConfig } from '@kubb/core'
+import type { Config } from '@kubb/core'
 
 export type Options = {
   /**
    * Kubb config without the Hooks.
    */
-  config?: Omit<UserConfig, 'hooks'>
+  config?: Omit<Config, 'hooks'>
 }
